+++ CHANGELOG.md
@@ -5,10 +5,12 @@
 
 - Cache `col_blocks` value in struct ([#66])
 - Add optional [Serde][serde] framework support ([#67])
+- Memory slice access for blocks ([#68])
 
 [#66]: https://github.com/gunvirranu/block-grid/pull/66
 [serde]: https://crates.io/crates/serde
 [#67]: https://github.com/gunvirranu/block-grid/pull/67
+[#68]: https://github.com/gunvirranu/block-grid/pull/68
 
 ## 0.1.1 - 2021-02-01